refactor: 모듈 통합으로 Redis 관련 패키지 구조 변경
