+++ issue-coupon/src/main/java/org/example/issuecoupon/config/RedisScriptConfig.java
@@ -1,4 +1,4 @@
-package org.example.redis.config;
+package org.example.issuecoupon.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
