+++ issue-coupon/src/main/java/org/example/issuecoupon/service/cache/cache/CouponIssueCacheStore.java
@@ -1,8 +1,6 @@
-package org.example.redis.service.port;
+package org.example.issuecoupon.service.cache.cache;
 
 public interface CouponIssueCacheStore {
 
     Long checkCouponIssueAvailability(String couponIssueRequestKey, String maxQuantity, String userId);
-
-    Long getIssuedCouponUserCount(String issuedCountKey);
 }
